# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation API E-Clinic",
        "title": "E-Clinic",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/e-clinic/v1",
    "paths": {
        "/poli": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Data Master Poli",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Management Poli"
                ],
                "summary": "Data Master Poli",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseRequestPagination"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Data Tambah Poli",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Management Poli"
                ],
                "summary": "Data Tambah Poli",
                "parameters": [
                    {
                        "description": "Create user body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/basemodel.RequestInsertBaseModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    }
                }
            }
        },
        "/poli/{id}": {
            "put": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Data Edit Poli",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Management Poli"
                ],
                "summary": "Data Edit Poli",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Poli ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create user body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/basemodel.RequestUpdateBaseModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Data Delete Poli",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Management Poli"
                ],
                "summary": "Data Delete Poli",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Poli ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "basemodel.RequestInsertBaseModel": {
            "type": "object",
            "properties": {
                "antreanCode": {
                    "type": "string"
                },
                "codeName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "basemodel.RequestUpdateBaseModel": {
            "type": "object",
            "properties": {
                "antreanCode": {
                    "type": "string"
                },
                "codeName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "basemodel.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "basemodel.ResponseRequestPagination": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "rows": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}